Add font category support for gx-edit component
Font category is special and rather different to other properties as it
can also be set through CSS (this is how GeneXus will define it, through
the CSS of the theme object).

To support this, a helper function was created called cssVariablesWatcher
that sets up a MutatioObserver to monitor changes on the style and class
attribute of an element, keeping the property in sync with its
corresponding CSS variable.

If the CSS variable has a value, it will update the property. If the class
or style attribute are changed, the property value is re-read and if it
changed, the property is updated, firing a re-render of the component.