Updated the expected result
Corrected the expected result.
Recently, the component form-field-render.tsx got a small update due to some labels were not placing correctly when using gx-edit. One of the label position corrected was the "bottom", in which case, it got updated to display with class "flex-column-reverse" not "flex-column". This test expected "flex-column", so that was the correction