Test fixes ncamera v1.1 (#210)
* Updated the expected result

Due to the recent changes in the gauge, the circle gauge rotations got a fix by 90 deg. Additionally, the test got a delay for 1s due to the "transition: $transitionConfig" property in the .circularMarker selector.

* Updated the expected result

Corrected the expected result.
Recently, the component form-field-render.tsx got a small update due to some labels were not placing correctly when using gx-edit. One of the label position corrected was the "bottom", in which case, it got updated to display with class "flex-column-reverse" not "flex-column". This test expected "flex-column", so that was the correction

* Updated the expected result

Probably, in the future it will have to be more rigorously corroborated if null was the expected result and not 32px. At this time, changing line 146 of image.tsx breaks other tests.