Test fixes ncamera v1.2 (#212)
* Fixed the false positive test and uncomented/improved another test

Both test were creating a gx-message to later corroborate if the gx-message was dismissed. In the first test, the condition to check the dismissed message was wait a certain duration and, in the second test, the condition was to click in a button.
Both tests got improved the condition to check if the gx-message is dismissed. The new condition is if the gx-message don't apear anymore in the HTML, since when a message is dismissed, it disapear from the HTML content.

* Fixed the false positives in the test

Due to it was missing another synchronization, the test was giving different results through different execution. Now it always gives the expected result.