Improved transitions in line gauge type
`lineCurrentValuePosition` and `lineIndicatorPosition` States were added to removed the style setters and remplace them for CSS class setters.

`.center-align` class is used for
   `margin-left: var(--percentage, 0)`
   `transform: translateX(-50%)`

`.right-align` class is used for
   `margin-left: 100%`
   `transform: translateX(-100%)`

`.left-align` class is the default class and does not need to be set in the HTML.

Also, starting to use `$transition-delay` variable on all available transitions.