Fixes for errors that occured when the properties `value` and `showValue` changes its values in runtime
These issues ocurred because the resizeObserver was not properly disconnected when the value of `showValue` or `value` properties changes at runtime.

Now, the resizeObserver will always disconnect when `showValue` or `value` toggles its value, and, after the end of the next rendering phase a proper resizeObserver will be connected to the control.