Add keys to children to avoid reuse of DOM elements
Because the toggle-button is always the initial child of its parent, when the back-button is added to the parent, it inherits the internal state of the toggle-button and the toggle-button is rebuilt as a new DOM element.
This happens because the toggle-button and back-button elements are practically identical (they differ only in the gx-icon type) and StencilJS optimizes these scenearios when the childs don't have keys.

More information: https://stenciljs.com/docs/templating-jsx#conditionals

issue: 93040