Added an aditional observer for gx-canvas control
This observer is used when: this.canvasFixedHeight == null

Let's assume the following scenario:
 - gx-canvas-cell with relative height (100%, for example)

 - gx-canvas with relative height (50%, for example)

 - The content of the gx-canvas-cell needs at least 200px of height space

 - The container of the gx-canvas has an initial height of 500px

In this case, if the container of the gx-canvas can decrease its height, when the gx-canvas has a height close to 200px, there will be a race condition in the updating of the height values of gx-canvas and gx-canvas-cell. This race condition will determine whether the gx-canvas sets a value for this.canvasFixedHeight or whether the gx-canvas keeps decreasing its height.

This observer is used to avoid the last case of the race condition.