`gx-canvas` refactor and other fixes (#252)
* Update .editorconfig to make "end_of_line" match in the Readme files

This prevents additional files from being changed when running the "npm run build" command.

* Update Readme files

* Update the interface of `gx-canvas-cell` control

These changes will allow us to have more control over the "autogrow" properties

* Removed previous implementation of `auto-grow`

* Added `minHeight` and `width` properties to the interface

* Implementation of autogrow in `gx-canvas` control when cells overflows

 - We use ResizeObservers to monitor size changes in the gx-canvas-cell controls

 - Every time the ResizeObserver interrupts, we check if there is a cell that has more height than the gx-canvas control. Otherwise, we check if the tallest gx-canvas-cell decreased its height when (in this case, we have to find a new "tallest" gx-canvas-cell to correct the height of the gx-canvas)

 - If the gx-canvas control never had a cell that exceeded its height, then: this.canvasFixedHeight == null.

 - When this.canvasFixedHeight turns stops being null, we fix the relative values of top, min-height and max-height in all the gx-canvas-cell.

* Added a subcontainer to support padding and border properties in the gx-canvas

Also, hide the overflow of gx-canvas-cell controls

issue: 94783

* Added an aditional observer for gx-canvas control

This observer is used when: this.canvasFixedHeight == null

Let's assume the following scenario:
 - gx-canvas-cell with relative height (100%, for example)

 - gx-canvas with relative height (50%, for example)

 - The content of the gx-canvas-cell needs at least 200px of height space

 - The container of the gx-canvas has an initial height of 500px

In this case, if the container of the gx-canvas can decrease its height, when the gx-canvas has a height close to 200px, there will be a race condition in the updating of the height values of gx-canvas and gx-canvas-cell. This race condition will determine whether the gx-canvas sets a value for this.canvasFixedHeight or whether the gx-canvas keeps decreasing its height.

This observer is used to avoid the last case of the race condition.

* Code reorder, update Readme and components.d.ts

* Hide content overflow in gx-canvas-cell

issue: 94810

* Add reflect to avoid generating default properties

* Added transition colors to `gx-canvas` control

issue: 95071

* Added layoutIsReady prop and removed setObserver() method

Instead of calling setObserver() method to notify when the gx-canvas control can set the auto-grow mechanism, we toggle the layoutIsReady property.

* Change default properties of gx-canvas-cell

* AutoGrow on gx-canvas when gx-canvas-cell is fixed

The auto-grow mechanism has been extended to allow auto-grow when any gx-canvas-cell has auto-grow = False (maxHeight != null).

* Added 2 border cases when the gx-canvas is resized

* Small improve for setCanvasMinHeight() function

If the gx-canvas is not visible, we only update the canvasFixedMinHeight variable and do not trigger any height adjustment, because we consider the gx-canvas not visible when this.element.clientHeight = 0.

* Update `canvasFixedMinHeight` variable the first time we fix the height of the gx-canvas

* Check layoutIsReady property when componentDidLoad

* Optimize the mutationObserver in gx-table-cell to reduce interruptions

The observer was subscribed to more attributes than the necessary. We just wanted to check value changes in the "hidden" attribute.

* Remove overflowMode from gx-canvas-cell interface

The gx-canvas control always has auto-grow = True and overflowMode is only used when auto-grow = False

* Add support for keep-space and collapse visibility

issue: 95124

* Improve gx-canvas observer disconnection

The gx-canvas observer will be disconnected the first time this.canvasFixedHeight != null

* Avoid extra calls for setCanvasAutoGrowMechanism()

In some scenarios, setCanvasAutoGrowMechanism() might be called twice. For example, when we have a gx-canvas inside of a gx-tab

* Check if componentDidLoad() was called before initializing the auto-grow mechanism

* Fix for autoGrow when the gx-canvas has padding

The current implementation now considers the top and bottom padding to implement the auto-grow mechanism.

* Code reuse using a function

* Fix for minHeight calculation when property has "calc"

Now, we add the absolute value of the "calc" property to the result.

* Update readme

* Change the default width to 100% in gx-canvas-cell