Fix auto-grow in gx-edit and gx-textblock controls when format="HTML"
Now, the gx-edit and gx-textblock controls have "auto-grow = True" when gx-table-cell container has the "auto-grow" attribute.
At this time, this attribute does not match any property in the gx-table-cell interface.
In the gx-canvas-cell, we do not use the auto-grow attribute, instead we use the max-height property which is always used when the inner control does not have AutoGrow.

issue: 96414