"runAlignmentTest" function for screenshot tests
"runAlignmentTest" is a function that can create up to 128 screenshot tests to check the alignment of controls. It receives the control we want to test, some metadata to identify the screenshot tests, and the configurations we want to test (check horizontal alignment, check vertical alignment and auto-grow values).

Once we have configured the function, it will create a screenshot unit test of the given control for each combination of container (gx-table and gx-canvas), cell height (80px and 100%), auto-grow values, horizontal alignment values and vertical alignment values.