Templates.ts file to facilitate the creation of controls for testing
This files provides a set of function to facilitate the creation of controls for the tests. Instead of writing the desired control string "by hand", we just call a function from this file with the properties we want to use in the control, and the function will return the desired string that models the behavior we want to test.