Add "updateScreenshots.mjs" script to update screenshot names
StencilJS does not provide us with an API to set the image name of each screenshots test (the screenshot name is an md5 hash  of the asociated test name), we have to update "manually" the names of the screeshots to have a good organization of the unit tests. Because ofthat, "updateScreenshots.mjs" is designed to give us with two commands to update the names of the screenshots.
 - The first command runs before the tests and is used to update the screenshot names from "custom name" to "original name" (this is necessary since StencilJS always compares the tests against the md5 hash, even if we update the "image" entries of the master.json).
 - The second command runs after the tests have been done and this command is used to update the names of the screenshots from "original name" to "custom name".

This script ("updateScreenshots.mjs") now needs two files to work; master.json and renames.json. master.json is automatically generated by StencilJS when we run the tests and renames.json is generated by us. renames.json is used to store the mapping between "custom name" and "original name" of screenshots.

To make it possible to create the rename.json file, we have to add a suffix to the test names: "image:<image name>".