gx-switch control refactoring (#261)
* Remove the old implementation of gx-switch

Also, the new implementation will not be split into a second file inside the "renders/bootstrap" folder. It will be in the switch.tsx file. The same will happen with gx-switch tests.

* New implementation for gx-switch control

 - gx-switch control design is based on Android and Unanimo designs.

 - Removed "checked" and "caption" properties.

 - 4 new properties were added that coincide with the switch control interface in GeneXus:
   - checkedCaption
   - checkedValue
   - unCheckedCaption
   - unCheckedValue

 - gx-switch now supports RTL.

* Added support for styling the gx-switch control

 - Added cssClass property to interface.

 - The cssClass property will be used to set the class attribute on the control.

* Add support for horizontal and vertical alignment

Also, start the refactoring for gx-form-field controls alignment. The old implementation still remains until we switch to the new one. The gx-switch control is the first component that uses the new implementation.