Improve screenshot unit tests (#263)
* Update test dependencies and remove dev vulnerabilities

* "runAlignmentTest" function for screenshot tests

"runAlignmentTest" is a function that can create up to 128 screenshot tests to check the alignment of controls. It receives the control we want to test, some metadata to identify the screenshot tests, and the configurations we want to test (check horizontal alignment, check vertical alignment and auto-grow values).

Once we have configured the function, it will create a screenshot unit test of the given control for each combination of container (gx-table and gx-canvas), cell height (80px and 100%), auto-grow values, horizontal alignment values and vertical alignment values.

* Templates.ts file to facilitate the creation of controls for testing

This files provides a set of function to facilitate the creation of controls for the tests. Instead of writing the desired control string "by hand", we just call a function from this file with the properties we want to use in the control, and the function will return the desired string that models the behavior we want to test.

* Update tsconfig.json to transpile the "tests" folder

* Add "updateScreenshots.mjs" script to update screenshot names

StencilJS does not provide us with an API to set the image name of each screenshots test (the screenshot name is an md5 hash  of the asociated test name), we have to update "manually" the names of the screeshots to have a good organization of the unit tests. Because ofthat, "updateScreenshots.mjs" is designed to give us with two commands to update the names of the screenshots.
 - The first command runs before the tests and is used to update the screenshot names from "custom name" to "original name" (this is necessary since StencilJS always compares the tests against the md5 hash, even if we update the "image" entries of the master.json).
 - The second command runs after the tests have been done and this command is used to update the names of the screenshots from "original name" to "custom name".

This script ("updateScreenshots.mjs") now needs two files to work; master.json and renames.json. master.json is automatically generated by StencilJS when we run the tests and renames.json is generated by us. renames.json is used to store the mapping between "custom name" and "original name" of screenshots.

To make it possible to create the rename.json file, we have to add a suffix to the test names: "image:<image name>".

* Tracking master.json and renames.json files

* Add screenshot unit tests for gx-checkbox alignment

* Add screenshot unit tests for gx-button alignment

* Add screenshot unit tests for gx-select alignment

* Add screenshot unit tests for gx-textblock alignment

* Add screenshot unit tests for gx-edit alignment

* Add screenshot unit tests for gx-radio-group alignment

* Fix default vertical alignment for gx-radio-group with direction="horizontal" and readonly="true"

* Fix gx-edit tests

* Temporarily comment out failed tests in gx-rating

Also, remove old gx-image alignment tests.