Many fixes (#273)
* Fix for lottie animations

 - gx-loading will now  take up space in its container

 - Changed the way that gx-loading is centered (simplified the CSS to center the gx-loading on grids).

 - When displayed a slot animation, the gx-loading will have a min-height of var(--height).

* Reduce bundle size by removing repeated imported styles

The styles on grid-base.scss are  imported multiple times, so we need to keep CSS in the file to a minimum size. In this case, part of the gx-grid-empty-indicator's styles were set on the grid-base.scss, so we moved them to their corresponing file.

* Fix for "updateLottiePath()" implementation

Introduced a new CSS variable ("--gx-animation-type") to implement some border cases that can not be achieved with a single variable. For example, having two clases, one with "gx-animation-type: lottie" and one with "gx-lottie-path: gx-file(...)" did not work with the old implementation.

* Remove nested selectors because control has Shadow DOM

Also, update "grow" animation to start using opacity instead of filter opacity

* Fix for targets (asides) being closed on wrong use cases

 - The targets will only close when the mask is clicked.

issue: 97413

* Improve gx-progress-bar's loading animation

* Fix for gx-image-picker's upload button not updating when the image is changed

 - Since Angular will not upload the image to the server when an image is changed, we no longer have to wait to update the state of the gx-image-picker's button.

 - Also, fix for add multiple times the "load" event to the reader.

 - Improve image load animation.

issue: 97634

* Add "scrollWheelZoom" property to enable/disable mouse wheel zoom