Fix for inverse loading when the grid is disabled but adds new items
In chat scenarios, the grid is disabled most of the time (it means, it loaded all its items), but since new items can be added dynamically, the resize observer and scroll handler should be set.

Added an additional condition in the scroll handler to not retrieve more items if the grid is fully loaded.