Fix gx-button's text-overflow in gx-tables with "overflow:visible"
When the gx-table has overflow: visible (Overflow Style: Show), the gx-button must limit its width to implement the text-overflow CSS property ("...").

To achieve this, the gx-button's content will have "position: absolute". For this to be posible, a couple of divs had to be added to correct the padding and implement the absolute positioning.

issue: 99173