Fix for auto-complete in the gx-edit control
We need to detect when the gx-edit value is auto completed to style the whole component, otherwise, only the input will be styled, leaving some unstyled parts in the gx-edit background.
To do this, an animation is set up to detect with JS when the auto complete triggers that animation.

issue: 100303