gx-maop refactoring
Remove the maxZoom property and the method it uses. Improve fitBounds() method documentation Set the maximum zoom of the map in the DidLoad component, using the maximum zoom constant defined For the selectionLayer in cssClass. Remove the css property (selectionTargetImageClass) that was used in the gx-map to implement this. Remove the selectionTargetImageSrc property.