Adjust the relative size of Flex Grid cells
When the Flex Grid has cells with relative size (width or height in percentage values), the cells must compute their size based on the size of the Flex Grid.

To achieve this, a resize observer depending on the scroll direction is configured to keep track of the Flex Grid size. The watcher is throttled with requestAnimationFrame to minimze interruptions.

issue: 101134