Optimize bundles and add more TypeScript constraints (#303)
* Split the grid bundle

The grids will no longer be in the same bundle. Each grid will be separated.

The infinite scroll controls will be in the same bundle. The same goes for the grid-smart-cell and grid-empty-indicator controls.

* Add the gx-loading and gx-lottie controls to the main bundle

Since lotties are supported in Angular applications, the gx-lottie's JS must be preloaded to display the lotties when they are needed on page load (Launch Screen Animation Class).

* Reorder bundles by name

* Add more TypeScript constraints

Now, all variables and parameters must have an explicit type.

* Fix missing type definition errors

* Improve gx-edit types

* Remove unused function

* Improve gx-query-viewer type definitions

* Update components.d.ts and fix for gx-map-circle

* Remove the gx-lottie control from the main bundle