Refactoring for gx-elevation property
The elevation shadow will no longer be hardcoded in the control's styles. Instead, it will be implemented through the box-shadow property that will come included in the control's class.

With this change, we can remove the most of the "classes.vars" styling in the controls' JS, which reduces blocking time.