Add support for Shadow DOM in the gx-edit control
To implement accesibility in the gx-edit control, the labelCaption property was added to the control. This property takes the same value as the labelCaption property of the gx-form-field parent.

Also, the label and the input will no longer be binded in the gx-edit, instead, when the gx-form-field's label is clicked, an event handler will click the gx-edit, setting focus in the input (or textarea) of the control.

issue: 102440